requirements.txt
# Core dependencies for QA Agent
# Python >= 3.9

# LangChain framework
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# OpenAI SDK for Doubao/Ark compatibility
openai>=1.40.0

# HTTP client
requests>=2.25.1

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=0.19.0

# CLI interface
typer>=0.9.0
rich>=13.0.0

# JSON/YAML processing
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.0

# Optional: Volcengine SDK (if needed as fallback)
# volcengine-python-sdk[ark]
